requirements.txt
openenv-core
fastapi
uvicorn
pydantic
openai
python-dotenv
numpy
pandas
google-genai
